# Conflicts:
#	spectra-generator/src/main/resources/testrunner.jar
#	trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/infrastructure/BufferedMap.java

+++ trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/infrastructure/comptrace/ReplaceableCloneableIterator.java
@@ -5,8 +5,8 @@ import se.de.hu_berlin.informatik.spectra.provider.tracecobertura.infrastructure
 
 public interface ReplaceableCloneableIterator<T> extends CloneableIterator<T> {
 
-	public T processNextAndReplaceWithResult(Function<T,T> function);
+    public T processNextAndReplaceWithResult(Function<T, T> function);
 
-	public void setToPosition(int i);
-	
+    public void setToPosition(long i);
+
 }
